Remove edit --fix option.
This sets up bug edit defaults to comment and close a bug after
pushing commits. Again, this is superceded by pulsebot.

This removes the only use of the fix_commits argument to edit_bug.

That in turn removes the only use of the status argument to
attach_commits, which in turn is the only caller of create_patch, so
we can remove the status argument to that, too.