Bundle up attachment requestee flags.
This changes the attachment flag information to be stored in a single
dictionary rather than a bunch of lists, and also stores the special
":me+" value that indicates a + separately. This will make it easier
to support more flags in the future.

I did not update the code in create_patch that actually sets the flags
in the request because that is all going to be completely removed
later.